+++ Video.h
@@ -11,29 +11,29 @@ using namespace std;
 class Video
 {
 private:
-    //Declaración de atributos 
+    // Declaración de atributos
     string id;
     string nombre;
     string duracion;
     string genero;
-    float calificacion;
+    double calificacion;
     string fecha;
 
 public:
-    //Constructor por defecto
+    // Constructor por defecto
     Video();
-    //Constructor que inicializa con los datos del Video
-    Video(string id, string nombre, string duracion, string genero, float calificacion, string fecha);
-    //setCalificación para cambiar la calificación
+    // Constructor que inicializa con los datos del Video
+    Video(string id, string nombre, string duracion, string genero, double calificacion, string fecha);
+    // setCalificación para cambiar la calificación
     void setCalificacion(float);
-    //get's devuelven un dato específico
+    // get's devuelven un dato específico
     string getId();
     string getNombre();
     string getDuracion();
     string getGenero();
-    float getCalificacion();
+    double getCalificacion();
     string getFecha();
-    //Devuelve los datos con formato
+    // Devuelve los datos con formato
     virtual void getInfo();
 };
